+++ CHANGELOG.rst
@@ -8,6 +8,11 @@ Notable changes to this project will be documented in this file.
 Upcoming
 ********
 
+Added
+=====
+
+- Add "No Kernel" to kernel selection list
+
 Fixed
 =====